port domain_simplify, update the expected outputs in the testsuite.
This is an inefficient implementation of domain_simplify.  For
matching the PolyLib implementation the code to detect redundant
constraints should be more accurate.  One solution could be to use the
GLPK solver that can be bound in with PPL to minimize the constraint
systems.  Until this happens, the output of the testsuite is modified
to reflect this new inefficient implementation of domain_simplify.