Fix compilation with -Wc++-compat warning.
Patch from Ian Lance Taylor <iant@google.com>

GCC, which now uses CLooG, aims to be written in the common subset of C
and C++.  To that end I have been enhancing GCC's -Wc++-compat warning.
A warning I committed yesterday triggers on the CLooG clast.h header
file.  This is http://gcc.gnu.org/PR40103 .

The problem is that some enum constants are defined inside of a struct.
In C++ that requires referring to the constants using the struct name as
a qualifier.